test: ポリシーダブル抽出とテストスイート修復 (#591)
TestRoomAccessService / TestIdentity を独立ファイル化し VulnerabilityTest を修復。
events API テストに start/end クエリを追加、AuditLogServiceTest の deprecation 修正。
